requirements.txt
Django==1.6.1
django-crispy-forms==1.4.0
South==0.8.4
dj-database-url==0.2.2
dj-static
gunicorn==18.0
psycopg2==2.5.1
